requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.6
django-autoslug==1.9.9
django-browser-reload==1.15.0
django-ckeditor-5==0.2.13
django-embed-video==1.4.10
django-etc==1.4.0
django-extensions==3.2.3
django-hitcount==1.3.5
django-recaptcha==4.0.0
django-taggit==5.0.1
idna==3.7
pillow==10.4.0
requests==2.32.3
sqlparse==0.5.0
urllib3==2.2.2
